Rename P2P utils file to miscellaneous
