+++ elements/lisk-p2p/src/peer_directory/tried_peers.ts
@@ -17,12 +17,12 @@ import {
 	DEFAULT_TRIED_PEER_BUCKET_SIZE,
 	DEFAULT_TRIED_PEER_LIST_SIZE,
 } from '..';
-import { P2PDiscoveredPeerInfo, P2PPeerInfo } from '../utils/types';
 import {
 	constructPeerIdFromPeerInfo,
 	getBucket,
 	PEER_TYPE,
-} from '../utils/utils';
+} from '../utils/miscellaneous';
+import { P2PDiscoveredPeerInfo, P2PPeerInfo } from '../utils/types';
 
 export interface TriedPeerConfig {
 	readonly triedPeerBucketCount?: number;
