+++ elements/lisk-p2p/src/utils/miscellaneous.ts
